+++ src/app/admin/honeymoon/guide/page.tsx
@@ -1,0 +1,8 @@
+'use client';
+
+import GuideTab from '../GuideTab';
+import { useHoneymoonApi } from '../HoneymoonContext';
+
+export default function HoneymoonGuidePage() {
+    return <GuideTab api={useHoneymoonApi()} />;
+}