+++ src/app/admin/honeymoon/itinerary/page.tsx
@@ -1,0 +1,8 @@
+'use client';
+
+import ItineraryTab from '../ItineraryTab';
+import { useHoneymoonApi } from '../HoneymoonContext';
+
+export default function HoneymoonItineraryPage() {
+    return <ItineraryTab api={useHoneymoonApi()} />;
+}