+++ src/app/admin/honeymoon/places/page.tsx
@@ -1,0 +1,8 @@
+'use client';
+
+import PlacesTab from '../PlacesTab';
+import { useHoneymoonApi } from '../HoneymoonContext';
+
+export default function HoneymoonPlacesPage() {
+    return <PlacesTab api={useHoneymoonApi()} />;
+}