+++ src/app/admin/honeymoon/settings/page.tsx
@@ -1,0 +1,8 @@
+'use client';
+
+import SettingsTab from '../SettingsTab';
+import { useHoneymoonApi } from '../HoneymoonContext';
+
+export default function HoneymoonSettingsPage() {
+    return <SettingsTab api={useHoneymoonApi()} />;
+}